Extract metadata types to shared gitWorkMetadata module
